Работающий подсчет очков, инфа 100%

+++ Assets/Scripts/GameData.cs
@@ -5,10 +5,10 @@ public class GameData : MonoBehaviour
     private static int playerMax = 4;
     private static int playerMin = 2;
 
-    private static float maxCountBal = 2f;
+    private static int maxCountBal = 2;
 
     public static int PlayerCount { get; set; }
     public static int PlayerMax => playerMax;
     public static int PlayerMin => playerMin;
-    public static float MaxCountBal => maxCountBal;
+    public static int MaxCountBal => maxCountBal;
 }
